Extract ImportAssets out of auto_import
